# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QuantStock API — NSE/BSE stock scoring engine
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface StockSearchResult {
  ticker: string;
  companyName: string;
  exchange: string;
}

export interface Stock {
  ticker: string;
  companyName: string;
  /** @nullable */
  sector: string | null;
  /** @nullable */
  industry: string | null;
  /** @nullable */
  marketCap: number | null;
  currency: string;
  /** @nullable */
  price: number | null;
  /** @nullable */
  changePercent?: number | null;
  lastUpdated: string;
}

export interface CategoryScores {
  valuation: number;
  financialHealth: number;
  profitability: number;
  growth: number;
  risk: number;
  momentum: number;
}

export type FactorDirection = typeof FactorDirection[keyof typeof FactorDirection];


export const FactorDirection = {
  positive: 'positive',
  negative: 'negative',
} as const;

export interface Factor {
  factor: string;
  direction: FactorDirection;
  weight: number;
}

export interface ScoreExplanation {
  reasons: string[];
  risks: string[];
}

export type ScoreSummaryRecommendation = typeof ScoreSummaryRecommendation[keyof typeof ScoreSummaryRecommendation];


export const ScoreSummaryRecommendation = {
  strong_buy: 'strong_buy',
  buy: 'buy',
  hold: 'hold',
  reduce: 'reduce',
  sell: 'sell',
  strong_short: 'strong_short',
} as const;

export interface ScoreSummary {
  ticker: string;
  overallScore: number;
  recommendation: ScoreSummaryRecommendation;
  confidence: number;
  /** @nullable */
  fairValueEstimate?: number | null;
  categoryScores: CategoryScores;
  topFactors: Factor[];
  riskFlags: string[];
  explanation: ScoreExplanation;
  computedAt: string;
}

export interface MetricDetail {
  key: string;
  label: string;
  /** @nullable */
  value: number | null;
  /** @nullable */
  sectorBenchmark?: number | null;
  score: number;
  higherIsBetter: boolean;
}

export interface CategoryDetail {
  category: string;
  weight: number;
  score: number;
  metrics: MetricDetail[];
}

export interface DcfAssumptions {
  wacc: number;
  terminalGrowthRate: number;
  projectionYears: number;
}

export type ScoreDetailsRecommendation = typeof ScoreDetailsRecommendation[keyof typeof ScoreDetailsRecommendation];


export const ScoreDetailsRecommendation = {
  strong_buy: 'strong_buy',
  buy: 'buy',
  hold: 'hold',
  reduce: 'reduce',
  sell: 'sell',
  strong_short: 'strong_short',
} as const;

export interface ScoreDetails {
  ticker: string;
  overallScore: number;
  recommendation: ScoreDetailsRecommendation;
  confidence: number;
  /** @nullable */
  fairValueEstimate?: number | null;
  categoryScores: CategoryScores;
  categories: CategoryDetail[];
  topFactors: Factor[];
  riskFlags: string[];
  explanation: ScoreExplanation;
  dcfAssumptions: DcfAssumptions;
  computedAt: string;
}

export interface PricePoint {
  date: string;
  close: number;
  volume: number;
}

export interface HistoryResponse {
  ticker: string;
  range: string;
  points: PricePoint[];
}

export interface ScoringConfigInput {
  /**
     * Sector name this override applies to, or null for the platform default
     * @nullable
     */
  sector: string | null;
  valuationWeight: number;
  financialHealthWeight: number;
  profitabilityWeight: number;
  growthWeight: number;
  riskWeight: number;
  momentumWeight: number;
}

export interface ScoringConfig {
  id: number;
  /** @nullable */
  sector: string | null;
  valuationWeight: number;
  financialHealthWeight: number;
  profitabilityWeight: number;
  growthWeight: number;
  riskWeight: number;
  momentumWeight: number;
  updatedAt: string;
}

export type SearchStocksParams = {
/**
 * @minLength 1
 */
q: string;
};