# lib/api-zod/src/generated/types/factor.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QuantStock API — NSE/BSE stock scoring engine
 * OpenAPI spec version: 0.1.0
 */
import type { FactorDirection } from './factorDirection';

export interface Factor {
  factor: string;
  direction: FactorDirection;
  weight: number;
}