# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QuantStock API — NSE/BSE stock scoring engine
 * OpenAPI spec version: 0.1.0
 */
import type { PricePoint } from './pricePoint';

export interface HistoryResponse {
  ticker: string;
  range: string;
  points: PricePoint[];
}