# lib/api-zod/src/generated/types/metricDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QuantStock API — NSE/BSE stock scoring engine
 * OpenAPI spec version: 0.1.0
 */

export interface MetricDetail {
  key: string;
  label: string;
  /** @nullable */
  value: number | null;
  /** @nullable */
  sectorBenchmark?: number | null;
  score: number;
  higherIsBetter: boolean;
}