# lib/api-zod/src/generated/types/pricePoint.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QuantStock API — NSE/BSE stock scoring engine
 * OpenAPI spec version: 0.1.0
 */

export interface PricePoint {
  date: Date;
  close: number;
  volume: number;
}